// src/_enums/regionsId.enum.ts
enum RegionsIdEnum {
  /** red, blue, green, yellow, fireRed, leafGreen */
  kanto = 'kanto',
  /** gold, silver, crystal */
  johto_1 = 'johto_1',
  /** heartGold, soulSilver */
  johto_2 = 'johto_2',
  /** ruby, sapphire, emerald */
  hoenn_1 = 'hoenn_1',
  /** omega_ruby, alpha_sapphire */
  hoenn_2 = 'hoenn_2',
  /** diamond, pearl */
  sinnoh_1 = 'sinnoh_1',
  /** platinium */
  sinnoh_2 = 'sinnoh_2',
  /** black, white */
  unova_1 = 'unova_1',
  /** black_2, white_2 */
  unova_2 = 'unova_2',
  /** x, y */
  kalos = 'kalos',
  /** sun, moon */
  aloha_1 = 'aloha_1',
  /** ultra_sun, ultra_moon */
  aloha_2 = 'aloha_2',
  /** sword, shield */
  galar = 'galar'
}

export default RegionsIdEnum